fix: circular dependency and uncomment runtime extrinsic
